Fixes the issue created on the Playground after Interpreters PR (#1553)
* WIP making PG functional again

* Code clean

* Add newly deployed addresses on Kovan

* Fix bug in truffle config syntax

* Add timeout back to view fn checking logic

* Remove debug logs

+++ packages/contracts/contracts/interpreters/ETHInterpreter.sol
@@ -25,11 +25,11 @@ contract ETHInterpreter is Interpreter {
     external
   {
 
-    ETHTransfer[] memory transfers =
-      abi.decode(input, (ETHTransfer[]));
+    ETHTransfer[] memory transfers = abi.decode(input, (ETHTransfer[]));
+
     uint256 limitRemaining = abi.decode(params, (Param)).limit;
 
-    for (uint i=0; i<transfers.length; i++) {
+    for (uint256 i = 0; i < transfers.length; i++) {
       address payable to = address(uint160(transfers[i].to));
       uint256 amount = transfers[i].amount;
